refactor: move scoring pipeline from core to engine layer
